Allow clients to explicitly derive generic Composites
I have a project where I have case classes which get used in doobie queries in
multiple different files. The case classes need assorted Meta instances to build
a Composite. Currently that means that each file that uses the case classes with
doobie has to manage making sure that the right Metas are in scope. Essentially
it's leaking the implementation into the other layers of the code.

It would be nice to define the Composite once for each case class in a companion
object, much like is done with circe. However, the following seems to compile
recursively and results in a null pointer exception:

    case class MyThing(...)
    object MyThing {
      implicit val composite: Composite[MyThing] = implicitly[Composite[MyThing]]
    }

The implicit "generic" function is not caller friendly, as you have to specify
the G hlist.

This commit adds the ability to do the following instead:

    case class MyThing(...)
    object MyThing {
      implicit val composite: Composite[MyThing] = Composite.ForGeneric[MyThing].derive()
    }