be community build friendly
also use latest sbt. might as well

dbuild, the meta-build tool we use in the Scala community build, needs
`onLoad` for its own purposes, and gets confused when a build
overrides it like this.

(also, nbd, but I would also argue that it's maybe not that great an idea
for a open source build to alter sbt's default behavior like this,
contributors may be confused that things aren't working as they would
normally expect)