Remove a majority of calls to throw
`RaiseError` has been added to the `Free` definitions and is used to
define `raiseError` instead of throwing in a `delay`.

Replace `unsafeFromInt` methods with `fromIntF` which requires an `F`
that can raise an error.

Update extension attempt methods to raise an error instead of throwing.

Throws in `Get` remain, removing them would probably have a performance
impact. These exceptions will be handled in `primitive` as it now
catches any exceptions and raises them in the effect.

A `temap` method has been added to `Get` to conversion functions to fail
with an error message instead of having to throw an exception. Under the
hood this just throws a `InvalidValue` invariant exception.