Add doobie-munit package to integrate Doobie with the shiny new MUnit
MUnit is a the new Scala testing framework on the block. Seems reasonable for
Doobie to support it along with the other two it already supports.

I copied off `scalatest` because it is fairly similar to MUnit. Hopefully I
didn't forget to change anything after the copy and paste, particularly in the
documentation (the test is green, at least).