Add syntactic sugar transactRaw to ConnectionIOOps
`ConnectionIOOps` already has syntax for calling `.transact(xa)` on a query, which applies `xa.trans`, but there is no equivalent for `xa.rawTrans`.

`xa.rawTrans` is useful when one does not want to use the `Strategy`, namely when we want to keep the `autoCommit` setting of the underlying DB connection instead of having Doobie override it to `false`, which it does by default. In other words, [as the Doobie documentation puts it](https://typelevel.org/doobie/docs/14-Managing-Connections.html#about-transactors), "This can be useful in cases **where transactional handling is unsupported or undesired**".

Here, we add `transactRaw` as a new syntax in `ConnectionIOOps`, paralleling the existing `transact`.

See the discussion here:
- https://github.com/typelevel/doobie/issues/1244#issuecomment-2757961204